Command.java: Changed command to primaryCommand. Added String commandSpecifier. Changed ORDER_ARGUMENT to ORDER_SPECIFIER. Changed PIZZA_ARGUMENT to PIZZA_SPECIFIER. began fixing some index out of bounds exceptions, still need to resolve those, see todo. Fixed bug in the evaluateCommand so that the user can type exit and not be asked to enter an argument. implemented the list method. Added functionality to print out invalid commands.
Menu.java:
Changed the constructor to be a normal method that sets up the menu and adds a few pizza to the static pizzaMenu.

OrderManager.java:
Instantiated the activeOrders and storedOrders ArrayLists.

PizzApp.java:
Added a line that calls the createMenu method in Menu.java.
Added a line that prints out "Please enter your command -> ".
Added a line that prints a new line to the console.
Added a few comments.